tests/wpt/web-platform-tests/paint-timing/first-image-child.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: child ignores parent FCP</title>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="image"></div>

<script>
async_test(function (t) {
    const iframe = document.createElement('iframe');
    iframe.id = 'child-iframe';
    iframe.src = 'resources/subframe-sending-paint.html';
    document.body.appendChild(iframe);
    window.addEventListener('message', t.step_func(e => {
        // Child iframe should not have any paint-timing entries.
        assert_equals(e.data, '0');
        t.done();
    }));
    const img = document.createElement('IMG');
    img.src = 'resources/circles.png';
    img.onload = function() {
        function testPaintEntries() {
            const bufferedEntries = performance.getEntriesByType('paint');
            if (bufferedEntries.length < 2) {
                t.step_timeout(testPaintEntries, 20);
                return;
            }
            t.step(function() {
                assert_equals(bufferedEntries.length, 2, 'There should be two paint timing instances.');
                assert_equals(bufferedEntries[0].entryType, 'paint');
                assert_equals(bufferedEntries[0].name, 'first-paint');
                assert_equals(bufferedEntries[1].entryType, 'paint');
                assert_equals(bufferedEntries[1].name, 'first-contentful-paint');
                // Ask child iframe to send its paint-timing entries.
                document.getElementById('child-iframe').
                    contentWindow.postMessage('', '*');
            })
        }
        testPaintEntries();
    };
    document.getElementById('image').appendChild(img);
}, 'Child iframe ignores paint-timing events fired from parent image rendering.');
</script>
</body>
</html>
